Fix: Refactor Navbar component directory
