+++ src/api/getUsersMe.tsx
@@ -3,7 +3,7 @@ import { axiosInstance } from "./axiosInstance";
 
 export default async function getUsersMe() {
   const accessToken =
-    "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6Njg5LCJ0ZWFtSWQiOiIyLTkiLCJpYXQiOjE3MDY2NzgwMzEsImlzcyI6InNwLXRhc2tpZnkifQ.xTJzppjh39utbp7V6-yYsFFXYzDmDT4jFUxabGtVZlY";
+    "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6NzkyLCJ0ZWFtSWQiOiIyLTkiLCJpYXQiOjE3MDY4NTI4OTgsImlzcyI6InNwLXRhc2tpZnkifQ.T-nZJzh7UKvzXXZARLI4W1Nq8CB-491fx-N16fUWW-g";
 
   try {
     const res = await axiosInstance.get("users/me", {
