+++ src/components/commons/BaseContainer/BaseContainer.module.scss
@@ -6,7 +6,6 @@
     "sidebar navbar"
     "sidebar content";
   scrollbar-gutter: stable;
-  border-right: 0.1rem solid $gray-300;
 
   @include responsive("T") {
     grid-template-columns: 16.8rem 1fr;
