+++ src/components/commons/BaseContainer/Navbar/Navbar.module.scss
@@ -139,104 +139,4 @@
       }
     }
   }
-  &-user {
-    @include flexbox;
-    gap: 0.9rem;
-    position: relative;
-    border-left: 1px solid $gray-300;
-    padding-left: 3.2rem;
-    cursor: pointer;
-
-    @include responsive("T") {
-      padding-left: 2.4rem;
-    }
-
-    @include responsive("M") {
-      padding-left: 1.2rem;
-    }
-
-    &-circle {
-      position: relative;
-      width: 3.8rem;
-      height: 3.8rem;
-      border: 0.2rem solid $white;
-      border-radius: 50%;
-      background-color: #ffc85a;
-
-      &-nickname {
-        position: absolute;
-        top: 50%;
-        left: 50%;
-        transform: translate(-50%, -50%);
-        @include text-style(1.6, 600, $white);
-      }
-    }
-    &-name {
-      @include text-style(1.6, 500, $black-200);
-
-      @include responsive("M") {
-        display: none;
-      }
-    }
-
-    &-dropdown-menu {
-      position: absolute;
-      top: 5rem;
-      padding: 1rem 2rem;
-      background-color: white;
-      border: 1px solid #ddd;
-      z-index: 1;
-
-      @include responsive("M") {
-        left: auto;
-        right: 0;
-      }
-
-      &::before {
-        content: "";
-        @include pos-center-x(absolute);
-        top: -8px;
-
-        border-color: #ddd transparent;
-        border-style: solid;
-        border-width: 0 6px 8px 6.5px;
-
-        @include responsive("M") {
-          left: auto;
-          right: 0.5rem;
-        }
-      }
-
-      &::after {
-        content: "";
-        display: block;
-        left: 75px;
-        @include pos-center-x(absolute);
-        top: -7px;
-        transform: translateX(-50%);
-        border-color: #fff transparent;
-        border-style: solid;
-        border-width: 0 6px 8px 6.5px;
-
-        @include responsive("M") {
-          left: auto;
-          right: 0.5rem;
-        }
-      }
-
-      button {
-        width: 100%;
-        text-align: center;
-      }
-
-      .text {
-        @include text-style(1.4, 600, $black-200);
-        line-height: 3;
-        white-space: nowrap;
-        &:hover {
-          color: $purple;
-        }
-      }
-    }
-  }
 }
